+++ Core/DTOs/InventoryCounting/InventoryCountingContentResponse.cs
@@ -5,9 +5,9 @@ public class InventoryCountingContentResponse {
     public string  ItemName        { get; set; } = string.Empty;
     public int?    BinEntry        { get; set; }
     public string? BinCode         { get; set; }
-    public int     SystemQuantity  { get; set; }
-    public int     CountedQuantity { get; set; }
-    public int     Variance        { get; set; }
+    public decimal SystemQuantity  { get; set; }
+    public decimal CountedQuantity { get; set; }
+    public decimal Variance        { get; set; }
     public decimal SystemValue     { get; set; }
     public decimal CountedValue    { get; set; }
     public decimal VarianceValue   { get; set; }
