+++ Core/DTOs/PickList/PickListCheckItemRequest.cs
@@ -5,7 +5,7 @@ namespace Core.DTOs.PickList;
 public class PickListCheckItemRequest {
     public int PickListId { get; set; }
     public string ItemCode { get; set; } = string.Empty;
-    public int CheckedQuantity { get; set; }
+    public decimal CheckedQuantity { get; set; }
     public UnitType Unit { get; set; }
     public int? BinEntry { get; set; }
 }
