+++ Core/Interfaces/IPickListPackageEligibilityService.cs
@@ -3,6 +3,6 @@ using Core.Entities;
 namespace Core.Interfaces;
 
 public interface IPickListPackageEligibilityService {
-    bool CanPackageBeFullyPicked(List<PackageContent> packageContents, Dictionary<string, int> itemOpenQuantities);
-    bool ValidatePackageForPicking(List<PackageContent> packageContents, Dictionary<string, int> itemOpenQuantities, out string? errorMessage);
+    bool CanPackageBeFullyPicked(List<PackageContent> packageContents, Dictionary<string, decimal> itemOpenQuantities);
+    bool ValidatePackageForPicking(List<PackageContent> packageContents, Dictionary<string, decimal> itemOpenQuantities, out string? errorMessage);
 }
